requirements.txt
shiny
urllib3
pillow
